+++ s3_url.go
@@ -1,0 +1,11 @@
+// Copyright (c) 2025 Benjamin Borbe All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package s3utils
+
+type URL string
+
+func (u URL) String() string {
+	return string(u)
+}